+++ app/main.cpp
@@ -7,7 +7,6 @@
 #include "config/global.h"
 #include "config/local.h"
 #include "config/storage.h"
-#include "radio/catalog.h"
 #ifdef ENABLE_MPD_SUPPORT
   #include "mpd_client/entity.h"
   #include "mpd_client/song.h"
@@ -131,10 +130,6 @@ int main(int argc, char *argv[]) {
 
   Config::Global global_conf;
   Config::Local local_conf;
-
-  if (!global_conf.hasRadioStations()) {
-    global_conf.saveRadioStations(Radio::Catalog::builtin());
-  }
 
   load_locale(a, global_conf.language());